requirements.txt
attrs==23.2.0
aws-cdk-lib==2.147.3
aws-cdk.asset-awscli-v2==2.0.118
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.166
boto3==1.26.65
botocore==1.29.65
cattrs==22.2.0
constructs==10.1.242
exceptiongroup==1.1.0
jmespath==1.0.1
jsii==1.101.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.14
pytest==6.2.5